fetch/api/response/response-error-from-stream.html

<!doctype html>
<html>
  <head>
    <meta charset="utf-8">
    <title>Response Receives Propagated Error from ReadableStream</title>
    <script src="/resources/testharness.js"></script>
    <script src="/resources/testharnessreport.js"></script>
  </head>
  <body>
    <script>
      /*************************************************************************
       *
       *  These tests originate from issue https://github.com/whatwg/fetch/issues/676
       *
       ************************************************************************/
      const errorThrown = new Error('ReadableStream error');

      function newStreamWithStartError() {
        return new ReadableStream({
          start(controller) {
            // controller.close();
            // controller.error(Error('Test 1 Error'));
            controller.error(errorThrown);
          }
        })
      }

      function newStreamWithPullError() {
        return new ReadableStream({
          pull(controller) {
            // controller.close();
            // controller.error(Error('Test 1 Error'));
            controller.error(errorThrown);
          }
        })
      }

      promise_test(_ => {
        return newStreamWithStartError().getReader().read()
        .then(
          ok => assert_false(true, 'ReadableStream Promise should throw error and not resolve'),
          error => assert_equals(error, errorThrown, "ReadableStream error does not propagate to Promise")
        )
      }, "ReadableStreamDefaultReader Promise receives ReadableStream start() Error")

      promise_test(_ => {
        return newStreamWithPullError().getReader().read()
        .then(
          ok => assert_false(true, 'ReadableStream Promise should throw error and not resolve'),
          error => assert_equals(error, errorThrown, "ReadableStream error does not propagate to Promise")
        )
      }, "ReadableStreamDefaultReader Promise receives ReadableStream pull() Error")

      promise_test(_ => {
        return new Response(newStreamWithStartError())
        .text()
        .then(
          ok => assert_false(true, 'ReadableStream Promise should throw error and not resolve'),
          error => assert_equals(error, errorThrown, "ReadableStream error does not propagate to Response Promise rejection")
        )
      }, "Response Promise receives ReadableStream start() Error")

      promise_test(_ => {
        return new Response(newStreamWithPullError())
        .text()
        .then(
          ok => assert_false(true, 'ReadableStream Promise should throw error and not resolve'),
          error => assert_equals(error, errorThrown, "ReadableStream error does not propagate to Response Promise rejection")
        )
      }, "Response Promise receives ReadableStream pull() Error")
    </script>
  </body>
</html>
